requirements.txt
Django==1.8.4
djangorestframework==3.0.1
oauthlib==1.0.3
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
